changed int in uint

+++ src/TheThingsNetwork.cpp
@@ -96,7 +96,7 @@ void TheThingsNetwork::reset(bool adr) {
   sendCommand(str);
 }
 
-void TheThingsNetwork::onMessage(void (*cb)(const byte* payload, size_t length, int8_t port)) {
+void TheThingsNetwork::onMessage(void (*cb)(const byte* payload, size_t length, uint8_t port)) {
   this->messageCallback = cb;
 }
 
@@ -130,7 +130,7 @@ bool TheThingsNetwork::provision(const byte appEui[8], const byte appKey[16]) {
   return sendCommand(F("mac save"));
 }
 
-bool TheThingsNetwork::join(int8_t retries, int32_t retryDelay) {
+bool TheThingsNetwork::join(int8_t retries, uint32_t retryDelay) {
   configureChannels(this->sf, this->fsb);
   String devEui = readValue(F("sys get hweui"));
   String str = "";
@@ -163,13 +163,13 @@ bool TheThingsNetwork::join(int8_t retries, int32_t retryDelay) {
   return false;
 }
 
-bool TheThingsNetwork::join(const byte appEui[8], const byte appKey[16], int8_t retries, int32_t retryDelay) {
+bool TheThingsNetwork::join(const byte appEui[8], const byte appKey[16], int8_t retries, uint32_t retryDelay) {
   reset();
   provision(appEui, appKey);
   return join(retries, retryDelay);
 }
 
-int TheThingsNetwork::sendBytes(const byte* payload, size_t length, int8_t port, bool confirm) {
+int TheThingsNetwork::sendBytes(const byte* payload, size_t length, uint8_t port, bool confirm) {
   String str = "";
   str.concat(F("mac tx "));
   str.concat(confirm ? F("cnf ") : F("uncnf "));
@@ -193,8 +193,8 @@ int TheThingsNetwork::sendBytes(const byte* payload, size_t length, int8_t port,
     return 1;
   }
   if (response.startsWith(F("mac_rx"))) {
-    int8_t portEnds = response.indexOf(" ", 7);
-    int8_t downlinkPort = response.substring(7, portEnds).toInt();
+    uint8_t portEnds = response.indexOf(" ", 7);
+    uint8_t downlinkPort = response.substring(7, portEnds).toInt();
     String data = response.substring(portEnds + 1);
     size_t downlinkLength = data.length() / 2;
     byte downlink[64];
@@ -214,7 +214,7 @@ int TheThingsNetwork::sendBytes(const byte* payload, size_t length, int8_t port,
   return -10;
 }
 
-int TheThingsNetwork::poll(int8_t port, bool confirm) {
+int TheThingsNetwork::poll(uint8_t port, bool confirm) {
   byte payload[] = { 0x00 };
   return sendBytes(payload, 1, port, confirm);
 }
@@ -227,7 +227,7 @@ void TheThingsNetwork::fillAirtimeInfo() {
   this->info.cr = 0;
   this->info.de = 0;
 
-  int16_t i;
+  uint8_t i;
   String message = readValue(F("radio get sf"));
   for (i = 2; message[i] && i <= 3; i++) {
     this->info.sf = (this->info.sf + message[i] - 48) * 10;
@@ -257,7 +257,7 @@ void TheThingsNetwork::trackAirtime(size_t payloadSize) {
 
   float Tsym = pow(2, this->info.sf) / this->info.band;
   float Tpreamble = (this->info.ps + 4.25) * Tsym;
-  int16_t payLoadSymbNb = 8 + (max(ceil((8 * payloadSize - 4 * this->info.sf + 28 + 16 - 20 * this->info.header) / (4 * (this->info.sf - 2 * this->info.de))) * (this->info.cr + 4), 0));
+  uint16_t payLoadSymbNb = 8 + (max(ceil((8 * payloadSize - 4 * this->info.sf + 28 + 16 - 20 * this->info.header) / (4 * (this->info.sf - 2 * this->info.de))) * (this->info.cr + 4), 0));
   float Tpayload = payLoadSymbNb * Tsym;
   float Tpacket = Tpreamble + Tpayload;
   this->airtime = this->airtime + (Tpacket / 1000);
@@ -288,10 +288,10 @@ void TheThingsNetwork::showStatus() {
   debugPrintLn(F(" s"));
 }
 
-void TheThingsNetwork::configureEU868(int8_t sf) {
-  int8_t ch;
+void TheThingsNetwork::configureEU868(uint8_t sf) {
+  uint8_t ch;
   int8_t dr = -1;
-  int32_t freq = 867100000;
+  uint32_t freq = 867100000;
   String str = "";
 
   str.concat(F("mac set rx2 3 869525000"));
@@ -360,13 +360,13 @@ void TheThingsNetwork::configureEU868(int8_t sf) {
   }
 }
 
-void TheThingsNetwork::configureUS915(int8_t sf, int8_t fsb) {
-  int8_t ch;
+void TheThingsNetwork::configureUS915(uint8_t sf, uint8_t fsb) {
+  uint8_t ch;
   int8_t dr = -1;
   String str = "";
-  int8_t chLow = fsb > 0 ? (fsb - 1) * 8 : 0;
-  int8_t chHigh = fsb > 0 ? chLow + 7 : 71;
-  int8_t ch500 = fsb + 63;
+  uint8_t chLow = fsb > 0 ? (fsb - 1) * 8 : 0;
+  uint8_t chHigh = fsb > 0 ? chLow + 7 : 71;
+  uint8_t ch500 = fsb + 63;
 
   sendCommand(F("radio set freq 904200000"));
   str = "";
@@ -419,7 +419,7 @@ void TheThingsNetwork::configureUS915(int8_t sf, int8_t fsb) {
   }
 }
 
-void TheThingsNetwork::configureChannels(int8_t sf, int8_t fsb) {
+void TheThingsNetwork::configureChannels(uint8_t sf, uint8_t fsb) {
   switch (this->fp) {
     case TTN_FP_EU868:
       configureEU868(sf);
@@ -437,7 +437,7 @@ void TheThingsNetwork::configureChannels(int8_t sf, int8_t fsb) {
   sendCommand(retries);
 }
 
-TheThingsNetwork::TheThingsNetwork(Stream& modemStream, Stream& debugStream, ttn_fp_t fp, int8_t sf, int8_t fsb) {
+TheThingsNetwork::TheThingsNetwork(Stream& modemStream, Stream& debugStream, ttn_fp_t fp, uint8_t sf, uint8_t fsb) {
   this->debugStream = &debugStream;
   this->modemStream = &modemStream;
   this->fp = fp;
